Rework docker build scripts for OWRX+ to build nighlty with all receivers and demodulators.
Add variables and rework the docker files to allow building with
different name (i.e. not 'openwebrx') and upload to different docker hub
account.
Rework the build system, to safe bandwidth (not installing dependencies
with every script) and CPU. On my system this made the building of the
full image in 35 minutes, instead of 60+ minutes. This will allow us to
have a -nightly builds, without spending much CPU and BW on the server,
which will compile them.