Fix docker build in lxc environment
Docker will use fuse-overlayfs in my lxc container.
The problem is that when the file is moved, the mv command returns with
error, saying it's the same file and cannot be moved. This is because
under the hood, the file is actually the same. Easiest fix is to cp it.